Add <triton> MiniMessage tag for translations
Fixes #252

This tag, to be used as `<triton:translationKey>` might be used to build
arbitrary style systems, as suggested by the issue above. It works by
inserting the literal value of the translation where the tag is.

So, considering a translation of key `color.primary` with the content
`<red>`.
A translation with the content `[minimsg]<triton:color.primary>hello!`
will be interpreted as `[minimsg]<red>hello!`.