fix(spigot): scoreboard not shown when using async mode
While using the experimental ProtocolLib async mode in Triton,
scoreboards, regardless if they had placeholders or not, were not shown.
This was caused by the packets not arriving in the same order they were
sent, since the intercepted packets arrived afterwards.

This commit registers all scoreboard related packets to ensure the order
is left unchanged.

This was previously "fixed" in v3 in #299 by forcing scoreboard packets
to be handled in the sync thread, which reduces performance.

Fixes #251