refactor(velocity): handle language player without player instance
This commit makes it possible to create a VelocityLanguagePlayer from a
UUID only, without using Velocity's Player instance. This is necessary
for plugins like LimboAPI that do not trigger LoginEvent while in the
limbo phase. For that reason, when translating via PacketEvents, Triton
would attempt to create instances of the VelocityLanguagePlayer but fail
due to the missing player instance. This has now been fixed by removing
the assumption that the Player instance is always available.

Fixes #498